requirements.txt
crewai
openai
python-dotenv
requests
langchain
langchain_community
google-generativeai
flask